Add processing of entities on batch (#59)
* Add processing of entities on batch

* Fix

* Fix

* WIP

* suggest fixed

* vector directory removed

* Vectors are now a list of float

* Preparing release 0.0.23

* Back to development: 0.0.24

* Fix vectors on float repeated

* WIP

* ix tests

* Fix reader

+++ nucliadb_protos/rust/src/utils.rs
@@ -72,8 +72,8 @@ pub struct Vector {
     pub start_paragraph: i32,
     #[prost(int32, tag="4")]
     pub end_paragraph: i32,
-    #[prost(bytes="vec", tag="5")]
-    pub vector: ::prost::alloc::vec::Vec<u8>,
+    #[prost(float, repeated, tag="5")]
+    pub vector: ::prost::alloc::vec::Vec<f32>,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Vectors {
